User sign in with email and password
Added ability for users to login with email and password. It uses
authlogic for managing and creating the session as well as hashing and
salting the password. Also added a namespace for dashboard controllers.
This commit includes yet support files for specs which include
ActionView::Helpers and hide logs from Puma.